色々修正、いらんもん消したり ステージ初期化(コンストラクタ)追加、大きさを指定したら自動で並べてくれるように、ただ増やせば増やすほど描画パスが増えるので超重くなる …どやって治そう？

+++ Sources/sprite_Batch.cpp
@@ -37,44 +37,15 @@ sprite_Batch::sprite_Batch(const wchar_t* filename, size_t max_sprites, const ch
 	hr = device->CreateBuffer(&buffer_desc, &subresource_data, vertex_buffer.GetAddressOf());		// 作成するバッファ情報、作成するバッファの初期化情報、作成したバッファを保存するポインタ
 	_ASSERT_EXPR(SUCCEEDED(hr), hr_trace(hr));	// _ASSERT_EXPR：第一引数条件が満たされなければ第二引数のメッセージを表示する
 
-	//	// 入力レイアウトオブジェクトの生成
-	//D3D11_INPUT_ELEMENT_DESC input_element_desc[]
-	//{
-	//	{
-	//	"POSITION",						// セマンティクス名	HLSL側のシグネチャ(型や変数名の組み合わせの事？)の名前と一致させることで送信した頂点情報を受信することができる
-	//	0,								// セマンティクス番号 同名でも識別できるように番号を割り当てる。番号を変更することでHLSLで別の情報だと認識できる
-	//	DXGI_FORMAT_R32G32B32_FLOAT,	// フォーマット	R23G23B23は実質float3
-	//	0,								// 入力スロット番号	入寮レイアウトをどの入力スロットに対して反映されるかを指定する
-	//	D3D11_APPEND_ALIGNED_ELEMENT,	// 要素から先頭までのオフセット値	各データの配列先頭が何バイト離れているか。
-	//									// D3D11_APPEND_ALIGNED_ELEMENTを指定でオフセット値を自動計算 手計算ならフォーマットサイズを加算していく
-	//	D3D11_INPUT_PER_VERTEX_DATA,	// 入力データの種類	頂点データとインスタンスデータの２種類
-	//	0								// 繰り返し回数(頂点データの時は０)	上記でインスタンスデータを設定した場合に意味を持つ
-	//	},
-	//	{"COLOR"   ,0,DXGI_FORMAT_R32G32B32A32_FLOAT,0,D3D11_APPEND_ALIGNED_ELEMENT,D3D11_INPUT_PER_VERTEX_DATA,0},
-	//	{"TEXCOORD",0,DXGI_FORMAT_R32G32_FLOAT		,0,D3D11_APPEND_ALIGNED_ELEMENT,D3D11_INPUT_PER_VERTEX_DATA,0},
-	//};
-
-	//// 頂点シェーダーオブジェクトの生成
-	//create_vs_from_cso(vs_cso_name, vertex_shader.GetAddressOf(), input_layout.GetAddressOf(), input_element_desc, _countof(input_element_desc));
-
-	//// ピクセルシェーダオブジェクトの生成
-	//create_ps_from_cso(ps_cso_name, pixel_shader.GetAddressOf());
-
 	param.Pos     = DirectX::SimpleMath::Vector2(0.0f, 0.0f);
-	param.Size    = DirectX::SimpleMath::Vector2(texture2d_desc.Width, texture2d_desc.Height);
+	param.Size = DirectX::SimpleMath::Vector2(static_cast<float>(texture2d_desc.Width), static_cast<float>(texture2d_desc.Height));
 	param.TexPos  = DirectX::SimpleMath::Vector2(0.0f, 0.0f);
-	param.TexSize = DirectX::SimpleMath::Vector2(texture2d_desc.Width, texture2d_desc.Height);
+	param.TexSize = DirectX::SimpleMath::Vector2(static_cast<float>(texture2d_desc.Width), static_cast<float>(texture2d_desc.Height));
 	param.Angle   = 0.0f;
 	param.Color   = DirectX::SimpleMath::Vector4(1.0f, 1.0f, 1.0f, 1.0f);
 }
 
 sprite_Batch::~sprite_Batch() {
-	//vertex_shader->Release();
-	//pixel_shader->Release();
-	//input_layout->Release();
-	//vertex_buffer->Release();
-	//shader_resource_view->Release();
-
 	rerease_all_textures();
 }
